+++ src/gamslib/formatdetect/jsontypes.py
@@ -4,6 +4,8 @@ import json
 from enum import StrEnum
 from pathlib import Path
 
+from gamslib.formatdetect.formatinfo import SubType
+
 
 # These MIME Types (als returned be a detection tool are handled as JSON files.)
 # This is an extension to the MIMETYPES dict, as some mime types listed here, are not
@@ -16,25 +18,16 @@ JSON_MIME_TYPES = [
 ]
 
 
-class JSONType(StrEnum):
-    "Define contants for known json types."
-
-    JSON = "JSON"
-    JSONLD = "JSON-LD"
-    JSONSCHEMA = "JSON-Schema"
-    JSONL = "JSON Lines"
-
-
 # I commented out some entries which are not yet registered with IANA.
 MIMETYPES = {
-    JSONType.JSON: "application/json",
-    JSONType.JSONLD: "application/ld+json",
+    SubType.JSON: "application/json",
+    SubType.JSONLD: "application/ld+json",
     # The suggest mime type is application/schema+json, but it is not registered yet
-    # JSONTypes.JSONSCHEMA: "application/schema+json",
-    JSONType.JSONSCHEMA: "application/json",
+    # SubType.JSONSCHEMA: "application/schema+json",
+    SubType.JSONSCHEMA: "application/json",
     # The suggest mime type is application/jsonl, but it is not registered yet
-    # JSONTypes.JSONL: "application/jsonl"
-    JSONType.JSONL: "application/json",
+    # SubType.JSONL: "application/jsonl"
+    SubType.JSONL: "application/json",
 }
 
 
@@ -61,14 +54,14 @@ def is_jsonl(data: str) -> bool:
     return is_jsonl_
 
 
-def guess_json_format(file_to_validate: Path) -> JSONType:
+def guess_json_format(file_to_validate: Path) -> SubType:
     """Guess the format of a JSON file.
 
     Returns:
         One of the formats defined in jsontypes.JSONTypes:
     """
     if file_to_validate.suffix == ".jsonld":
-        return JSONType.JSONLD
+        return SubType.JSONLD
 
     try:
         with open(file_to_validate, "r", encoding="utf-8", newline="") as f:
@@ -79,24 +72,24 @@ def guess_json_format(file_to_validate: Path) -> JSONType:
                 and jsondata["$schema"]
                 == "https://json-schema.org/draft/2020-12/schema"
             ):
-                return JSONType.JSONSCHEMA
+                return SubType.JSONSCHEMA
 
             for key in jsondata:
                 if key in ["@context", "@id"]:
-                    return JSONType.JSONLD
+                    return SubType.JSONLD
     # if file contains jsonl context, parsing will fail
     except json.JSONDecodeError as exp:
         if is_jsonl(file_content):
-            return JSONType.JSONL
+            return SubType.JSONL
         raise exp from exp  # eg. invalid JSON
-    return JSONType.JSON
+    return SubType.JSON
 
 
-def get_format_info(filepath: Path, mime_type: str) -> tuple[str, str]:
+def get_format_info(filepath: Path, mime_type: str) -> tuple[str, SubType | None]:
     """Return a tuple with (possibly fixed) mimetype and subtype."""
-    subtype = ""
+    subtype = None
     json_type = guess_json_format(filepath)
     if json_type in MIMETYPES:
         mime_type = MIMETYPES[json_type]
-        subtype = json_type.value
+        subtype = json_type
     return mime_type, subtype