+++ tests/formatdetect/test_xmltypes.py
@@ -1,10 +1,10 @@
 import pytest
 from gamslib.formatdetect.xmltypes import is_xml_type
 from pathlib import Path
-from gamslib.formatdetect.xmltypes import guess_xml_subtype, XMLTypes
+from gamslib.formatdetect.xmltypes import guess_xml_subtype
 import warnings
 from gamslib.formatdetect.xmltypes import get_format_info
-from gamslib.formatdetect.formatinfo import FormatInfo
+from gamslib.formatdetect.formatinfo import FormatInfo, SubType
 
 def test_is_xml_type_known_mimetype():
     assert is_xml_type("application/xml") == True
@@ -25,7 +25,7 @@ def test_guess_xml_subtype_known_namespace(tmp_path):
     xml_file = tmp_path / "test_known_namespace.xml"
     xml_file.write_text(xml_content)
 
-    assert guess_xml_subtype(xml_file) == XMLTypes.ATOM
+    assert guess_xml_subtype(xml_file) == SubType.ATOM
 
 def test_guess_xml_subtype_unknown_namespace(tmp_path):
     xml_content = """<?xml version="1.0" encoding="UTF-8"?>
@@ -59,7 +59,7 @@ def test_get_format_info_known_namespace(tmp_path):
 
     mimetype, subtype = get_format_info(xml_file, "application/xml")
     assert mimetype == "application/atom+xml"
-    assert subtype == "ATOM"
+    assert subtype == SubType.ATOM
 
 def test_get_format_info_unknown_namespace(tmp_path):
     xml_content = """<?xml version="1.0" encoding="UTF-8"?>
@@ -71,7 +71,7 @@ def test_get_format_info_unknown_namespace(tmp_path):
     with pytest.warns(UserWarning):
         mimetype, subtype = get_format_info(xml_file, "application/xml")
         assert mimetype == "application/xml"
-        assert subtype == ""
+        assert subtype is None
 
 def test_get_format_info_no_namespace(tmp_path):
     xml_content = """<?xml version="1.0" encoding="UTF-8"?>
@@ -82,7 +82,7 @@ def test_get_format_info_no_namespace(tmp_path):
 
     mimetype, subtype = get_format_info(xml_file, "application/xml")
     assert mimetype == "application/xml"
-    assert subtype == ""
+    assert subtype is None